+++ docs/changelog.rst
@@ -12,6 +12,11 @@ New Features
 ============
 - Experimental waves build command for automatically re-running workflows which extend a parameter study (:issue:`279`,
   :merge:`285`). By `Kyle Brindley`_.
+- Add a custom study subcommand to the parameter study CLI (:issue:`276`, :merge:`289`). By `Kyle Brindley`_.
+
+Bug fixes
+=========
+- Accept any 2D array like for the Custom Study parameter generator (:issue:`276`, :merge:`289`). By `Kyle Brindley`_.
 
 Documentation
 =============
@@ -24,6 +29,10 @@ Enhancements
 ============
 - Provide the data downselection dictionary as a CLI argument rather than hardcoding it in ``plot_scatter.py``
   (:issue:`273`, :merge:`281`). By `Thomas Roberts`_.
+
+Internal Changes
+================
+- Test the parameter study CLI generator sub-commands (:issue:`276`, :merge:`289`). By `Kyle Brindley`_.
 
 ******************
 0.3.6 (2022-08-31)